build with "-Wall -Werror" and fix all the build failures this causes
This fixes several potential bugs.

I also got it building with clang++, although I currently have no plans to add
it to my travis-ci build matrix because I want to be a good citizen and avoid
spawning 16 build jobs on every push.

Also, it's weird that the CMP/PL and CMP/PZ bugs that got fixed were never
caught by sh4inst_test.  This may require further investigation.

+++ src/hw/sh4/sh4_inst.cpp
@@ -941,7 +941,7 @@ void Sh4::inst_unary_movt_gen(OpArgs inst) {
 // 0100nnnn00010001
 void Sh4::inst_unary_cmppz_gen(OpArgs inst) {
     reg.sr &= ~SR_FLAG_T_MASK;
-    uint32_t flag = (*gen_reg(inst.gen_reg)) >= 0;
+    uint32_t flag = int32_t(*gen_reg(inst.gen_reg)) >= 0;
 
     reg.sr |= flag << SR_FLAG_T_SHIFT;
 
@@ -952,7 +952,7 @@ void Sh4::inst_unary_cmppz_gen(OpArgs inst) {
 // 0100nnnn00010101
 void Sh4::inst_unary_cmppl_gen(OpArgs inst) {
     reg.sr &= ~SR_FLAG_T_MASK;
-    uint32_t flag = (*gen_reg(inst.gen_reg)) > 0;
+    uint32_t flag = int32_t(*gen_reg(inst.gen_reg)) > 0;
 
     reg.sr |= flag << SR_FLAG_T_SHIFT;
